:white_check_mark: 위/경도 기반 주변 꽃집 찾기 테스트 #LF1-290

+++ src/main/java/kr/bb/store/domain/store/handler/StoreReader.java
@@ -73,7 +73,7 @@ public class StoreReader {
                 .build();
     }
 
-    public StoresByLocationResponse getNearbyStores(Float lat, Float lon) {
+    public StoresByLocationResponse getNearbyStores(Double lat, Double lon) {
         List<StoreForMapResponse> nearbyStores = storeRepository.getNearbyStores(lat, lon, RADIUS_FOR_MAP);
         // TODO : 좋아요 여부 feign으로 받아와서 채우기
         return StoresByLocationResponse.builder()
